+++ zcash_primitives/src/transaction/builder.rs
@@ -4,8 +4,13 @@ use core::cmp::Ordering;
 use core::fmt;
 use rand::{CryptoRng, RngCore};
 
-use ::sapling::{Note, PaymentAddress, builder::SaplingMetadata};
-use ::transparent::{address::TransparentAddress, builder::TransparentBuilder, bundle::TxOut};
+use sapling::{Note, PaymentAddress, builder::SaplingMetadata};
+use transparent::{
+    address::TransparentAddress,
+    builder::TransparentBuilder,
+    bundle::TxOut,
+    coinbase::{self, MinerData},
+};
 use zcash_protocol::{
     consensus::{self, BlockHeight, BranchId, NetworkUpgrade, Parameters},
     memo::MemoBytes,
@@ -30,13 +35,13 @@ use {
         sighash::{SignableInput, signature_hash},
         txid::TxIdDigester,
     },
-    ::sapling::prover::{OutputProver, SpendProver},
-    ::transparent::builder::TransparentSigningSet,
     alloc::vec::Vec,
+    sapling::prover::{OutputProver, SpendProver},
+    transparent::builder::TransparentSigningSet,
 };
 
 #[cfg(feature = "transparent-inputs")]
-use {::transparent::builder::TransparentInputInfo, zcash_script::script};
+use {transparent::builder::TransparentInputInfo, zcash_script::script};
 
 #[cfg(not(feature = "transparent-inputs"))]
 use core::convert::Infallible;
@@ -53,7 +58,7 @@ use crate::{
     },
 };
 
-use super::components::sapling::zip212_enforcement;
+use super::{Coinbase, components::sapling::zip212_enforcement};
 
 /// Since Blossom activation, the default transaction expiry delta should be 40 blocks.
 /// <https://zips.z.cash/zip-0203#changes-for-blossom>
@@ -104,6 +109,8 @@ pub enum Error<FE> {
     /// The builder was constructed with a target height before NU5 activation, but an Orchard
     /// spend or output was added.
     OrchardBuilderNotAvailable,
+    /// An error occurred in constructing a coinbase transaction.
+    Coinbase(coinbase::Error),
     /// An error occurred in constructing the TZE parts of a transaction.
     #[cfg(zcash_unstable = "zfuture")]
     TzeBuild(tze::builder::Error),
@@ -135,6 +142,10 @@ impl<FE: fmt::Display> fmt::Display for Error<FE> {
                 f,
                 "Cannot create Orchard transactions without an Orchard anchor, or before NU5 activation"
             ),
+            Error::Coinbase(err) => write!(
+                f,
+                "An error occurred in constructing a coinbase transaction: {err}"
+            ),
             #[cfg(zcash_unstable = "zfuture")]
             Error::TzeBuild(err) => err.fmt(f),
         }
@@ -165,6 +176,12 @@ impl<FE> From<sapling::builder::Error> for Error<FE> {
 impl<FE> From<orchard::builder::SpendError> for Error<FE> {
     fn from(e: orchard::builder::SpendError) -> Self {
         Error::OrchardSpend(e)
+    }
+}
+
+impl<FE> From<coinbase::Error> for Error<FE> {
+    fn from(e: coinbase::Error) -> Self {
+        Error::Coinbase(e)
     }
 }
 
@@ -202,14 +219,16 @@ impl Progress {
     }
 }
 
-/// Rules for how the builder should be configured for each shielded pool.
-#[derive(Clone, Copy)]
+/// Rules for how the builder should be configured for each shielded pool or coinbase tx.
+#[derive(Clone)]
 pub enum BuildConfig {
     Standard {
         sapling_anchor: Option<sapling::Anchor>,
         orchard_anchor: Option<orchard::Anchor>,
     },
-    Coinbase,
+    Coinbase {
+        miner_data: MinerData,
+    },
 }
 
 impl BuildConfig {
@@ -221,7 +240,7 @@ impl BuildConfig {
             BuildConfig::Standard { sapling_anchor, .. } => sapling_anchor
                 .as_ref()
                 .map(|a| (sapling::builder::BundleType::DEFAULT, *a)),
-            BuildConfig::Coinbase => Some((
+            BuildConfig::Coinbase { .. } => Some((
                 sapling::builder::BundleType::Coinbase,
                 sapling::Anchor::empty_tree(),
             )),
@@ -236,11 +255,16 @@ impl BuildConfig {
             BuildConfig::Standard { orchard_anchor, .. } => orchard_anchor
                 .as_ref()
                 .map(|a| (orchard::builder::BundleType::DEFAULT, *a)),
-            BuildConfig::Coinbase => Some((
+            BuildConfig::Coinbase { .. } => Some((
                 orchard::builder::BundleType::Coinbase,
                 orchard::Anchor::empty_tree(),
             )),
         }
+    }
+
+    /// Returns `true` if this configuration is for building a coinbase transaction.
+    pub fn is_coinbase(&self) -> bool {
+        matches!(self, BuildConfig::Coinbase { .. })
     }
 }
 
@@ -386,11 +410,26 @@ impl<'a, P: consensus::Parameters> Builder<'a, P, ()> {
                 )
             });
 
+        // # Consensus Rules
+        //
+        // > [NU5 onward] The `nExpiryHeight` field of a coinbase transaction MUST be equal to its
+        // > block height.
+        //
+        // ## Notes
+        //
+        // We set the expiry height for coinbase txs to the block height regardless of the network
+        // upgrade.
+        let expiry_height = if build_config.is_coinbase() {
+            target_height
+        } else {
+            target_height + DEFAULT_TX_EXPIRY_DELTA
+        };
+
         Builder {
             params,
             build_config,
             target_height,
-            expiry_height: target_height + DEFAULT_TX_EXPIRY_DELTA,
+            expiry_height,
             #[cfg(all(
                 any(zcash_unstable = "nu7", zcash_unstable = "zfuture"),
                 feature = "zip-233"
@@ -461,7 +500,7 @@ impl<P: consensus::Parameters, U: sapling::builder::ProverProgress> Builder<'_, 
         &mut self,
         ovk: Option<orchard::keys::OutgoingViewingKey>,
         recipient: orchard::Address,
-        value: u64,
+        value: Zatoshis,
         memo: MemoBytes,
     ) -> Result<(), Error<FE>> {
         self.orchard_builder
@@ -470,7 +509,7 @@ impl<P: consensus::Parameters, U: sapling::builder::ProverProgress> Builder<'_, 
             .add_output(
                 ovk,
                 recipient,
-                orchard::value::NoteValue::from_raw(value),
+                orchard::value::NoteValue::from_raw(value.into()),
                 memo.into_bytes(),
             )
             .map_err(Error::OrchardRecipient)
@@ -692,8 +731,17 @@ impl<P: consensus::Parameters, U: sapling::builder::ProverProgress> Builder<'_, 
 
     /// Builds a transaction from the configured spends and outputs.
     ///
-    /// Upon success, returns a tuple containing the final transaction, and the
-    /// [`SaplingMetadata`] generated during the build process.
+    /// Upon success, returns a [`BuildResult`] containing:
+    ///
+    /// - the [final transaction],
+    /// - the [Sapling metadata], and
+    /// - the [Orchard metadata]
+    ///
+    /// generated during the build process.
+    ///
+    /// [Sapling metadata]: ::sapling::builder::SaplingMetadata
+    /// [Orchard metadata]: ::orchard::builder::BundleMetadata
+    /// [final transaction]: Transaction
     #[allow(clippy::too_many_arguments)]
     #[cfg(feature = "circuits")]
     pub fn build<R: RngCore + CryptoRng, SP: SpendProver, OP: OutputProver, FR: FeeRule>(
@@ -706,22 +754,176 @@ impl<P: consensus::Parameters, U: sapling::builder::ProverProgress> Builder<'_, 
         output_prover: &OP,
         fee_rule: &FR,
     ) -> Result<BuildResult, Error<FR::Error>> {
-        let fee = self.get_fee(fee_rule).map_err(Error::Fee)?;
-        self.build_internal(
-            transparent_signing_set,
-            sapling_extsks,
-            orchard_saks,
-            rng,
-            spend_prover,
-            output_prover,
-            fee,
-        )
+        if self.build_config.is_coinbase() {
+            self.build_coinbase(rng, spend_prover, output_prover)
+        } else {
+            let fee = self.get_fee(fee_rule).map_err(Error::Fee)?;
+
+            self.build_standard(
+                transparent_signing_set,
+                sapling_extsks,
+                orchard_saks,
+                rng,
+                spend_prover,
+                output_prover,
+                fee,
+            )
+        }
+    }
+
+    #[cfg(feature = "circuits")]
+    fn build_coinbase<R: RngCore + CryptoRng, SP: SpendProver, OP: OutputProver, FE>(
+        self,
+        mut rng: R,
+        spend_prover: &SP,
+        output_prover: &OP,
+    ) -> Result<BuildResult, Error<FE>> {
+        use super::Authorized;
+
+        let consensus_branch_id = BranchId::for_height(&self.params, self.target_height);
+
+        let miner_data = match self.build_config {
+            BuildConfig::Coinbase { miner_data } => miner_data,
+            _ => panic!(
+                "the build_coinbase method should not be called without a coinbase build config"
+            ),
+        };
+
+        let transparent_bundle = self
+            .transparent_builder
+            .build_coinbase(self.target_height, &miner_data)?;
+
+        let (sapling_bundle, sapling_meta) = match self
+            .sapling_builder
+            .and_then(|builder| {
+                builder
+                    .build::<SP, OP, _, _>(&[], &mut rng)
+                    .map_err(Error::SaplingBuild)
+                    .transpose()
+                    .map(|res| {
+                        res.map(|(bundle, sapling_meta)| {
+                            (
+                                bundle.create_proofs(
+                                    spend_prover,
+                                    output_prover,
+                                    &mut rng,
+                                    self._progress_notifier,
+                                ),
+                                sapling_meta,
+                            )
+                        })
+                    })
+            })
+            .transpose()?
+        {
+            Some((bundle, meta)) => (Some(bundle), meta),
+            None => (None, SaplingMetadata::empty()),
+        };
+
+        let (orchard_bundle, orchard_meta) = match self
+            .orchard_builder
+            .and_then(|builder| {
+                builder
+                    .build(&mut rng)
+                    .map_err(Error::OrchardBuild)
+                    .transpose()
+            })
+            .transpose()?
+        {
+            Some((bundle, meta)) => (Some(bundle), meta),
+            None => (None, orchard::builder::BundleMetadata::empty()),
+        };
+
+        let unauthed_tx: TransactionData<Coinbase> = TransactionData {
+            version: TxVersion::suggested_for_branch(consensus_branch_id),
+            consensus_branch_id,
+            // We don't use lock_time in coinbase transactions.
+            lock_time: 0,
+            expiry_height: self.expiry_height,
+            #[cfg(all(
+                any(zcash_unstable = "nu7", zcash_unstable = "zfuture"),
+                feature = "zip-233"
+            ))]
+            zip233_amount: self.zip233_amount,
+            transparent_bundle: Some(transparent_bundle.clone()),
+            // # Consensus
+            //
+            // > A coinbase transaction MUST NOT have any JoinSplit descriptions.
+            //
+            // > A coinbase transaction MUST NOT have any Spend descriptions.
+            //
+            // <https://zips.z.cash/protocol/protocol.pdf#txnconsensus>
+            sprout_bundle: None,
+            sapling_bundle,
+            orchard_bundle,
+            #[cfg(zcash_unstable = "zfuture")]
+            tze_bundle: None,
+        };
+
+        let transparent_bundle =
+            transparent_bundle.map_authorization(transparent::builder::Coinbase);
+
+        let shielded_sig_commitment = signature_hash(
+            &unauthed_tx,
+            &SignableInput::Shielded,
+            &unauthed_tx.digest(TxIdDigester),
+        );
+
+        let sapling_bundle = unauthed_tx
+            .sapling_bundle
+            .map(|b| b.apply_signatures(&mut rng, *shielded_sig_commitment.as_ref(), &[]))
+            .transpose()
+            .map_err(Error::SaplingBuild)?;
+
+        let orchard_bundle = unauthed_tx
+            .orchard_bundle
+            .map(|b| {
+                b.create_proof(&orchard::circuit::ProvingKey::build(), &mut rng)
+                    .and_then(|b| {
+                        b.apply_signatures(&mut rng, *shielded_sig_commitment.as_ref(), &[])
+                    })
+            })
+            .transpose()
+            .map_err(Error::OrchardBuild)?;
+
+        let authed_tx: TransactionData<Authorized> = TransactionData {
+            version: unauthed_tx.version,
+            consensus_branch_id: unauthed_tx.consensus_branch_id,
+            lock_time: unauthed_tx.lock_time,
+            expiry_height: unauthed_tx.expiry_height,
+            #[cfg(all(
+                any(zcash_unstable = "nu7", zcash_unstable = "zfuture"),
+                feature = "zip-233"
+            ))]
+            zip233_amount: self.zip233_amount,
+            transparent_bundle: Some(transparent_bundle),
+            sprout_bundle: unauthed_tx.sprout_bundle,
+            sapling_bundle,
+            orchard_bundle,
+            #[cfg(zcash_unstable = "zfuture")]
+            tze_bundle: None,
+        };
+
+        Ok(BuildResult {
+            transaction: authed_tx.freeze().unwrap(),
+            sapling_meta,
+            orchard_meta,
+        })
     }
 
     /// Builds a transaction from the configured spends and outputs.
     ///
-    /// Upon success, returns a tuple containing the final transaction, and the
-    /// [`SaplingMetadata`] generated during the build process.
+    /// Upon success, returns a [`BuildResult`] containing:
+    ///
+    /// - the [final transaction],
+    /// - the [Sapling metadata], and
+    /// - the [Orchard metadata]
+    ///
+    /// generated during the build process.
+    ///
+    /// [Sapling metadata]: ::sapling::builder::SaplingMetadata
+    /// [Orchard metadata]: ::orchard::builder::BundleMetadata
+    /// [final transaction]: Transaction
     #[cfg(zcash_unstable = "zfuture")]
     pub fn build_zfuture<
         R: RngCore + CryptoRng,
@@ -738,21 +940,26 @@ impl<P: consensus::Parameters, U: sapling::builder::ProverProgress> Builder<'_, 
         output_prover: &OP,
         fee_rule: &FR,
     ) -> Result<BuildResult, Error<FR::Error>> {
-        let fee = self.get_fee_zfuture(fee_rule).map_err(Error::Fee)?;
-        self.build_internal(
-            transparent_signing_set,
-            sapling_extsks,
-            orchard_saks,
-            rng,
-            spend_prover,
-            output_prover,
-            fee,
-        )
+        if self.build_config.is_coinbase() {
+            self.build_coinbase(rng, spend_prover, output_prover)
+        } else {
+            let fee = self.get_fee(fee_rule).map_err(Error::Fee)?;
+
+            self.build_standard(
+                transparent_signing_set,
+                sapling_extsks,
+                orchard_saks,
+                rng,
+                spend_prover,
+                output_prover,
+                fee,
+            )
+        }
     }
 
     #[allow(clippy::too_many_arguments)]
     #[cfg(feature = "circuits")]
-    fn build_internal<R: RngCore + CryptoRng, SP: SpendProver, OP: OutputProver, FE>(
+    fn build_standard<R: RngCore + CryptoRng, SP: SpendProver, OP: OutputProver, FE>(
         self,
         transparent_signing_set: &TransparentSigningSet,
         sapling_extsks: &[sapling::zip32::ExtendedSpendingKey],
@@ -763,9 +970,6 @@ impl<P: consensus::Parameters, U: sapling::builder::ProverProgress> Builder<'_, 
         fee: Zatoshis,
     ) -> Result<BuildResult, Error<FE>> {
         let consensus_branch_id = BranchId::for_height(&self.params, self.target_height);
-
-        // determine transaction version
-        let version = TxVersion::suggested_for_branch(consensus_branch_id);
 
         //
         // Consistency checks
@@ -834,7 +1038,7 @@ impl<P: consensus::Parameters, U: sapling::builder::ProverProgress> Builder<'_, 
         let (tze_bundle, tze_signers) = self.tze_builder.build();
 
         let unauthed_tx: TransactionData<Unauthorized> = TransactionData {
-            version,
+            version: TxVersion::suggested_for_branch(consensus_branch_id),
             consensus_branch_id: BranchId::for_height(&self.params, self.target_height),
             lock_time: 0,
             expiry_height: self.expiry_height,
@@ -1060,8 +1264,8 @@ mod testing {
     use rand::RngCore;
     use rand_core::CryptoRng;
 
-    use ::sapling::prover::mock::{MockOutputProver, MockSpendProver};
-    use ::transparent::builder::TransparentSigningSet;
+    use sapling::prover::mock::{MockOutputProver, MockSpendProver};
+    use transparent::builder::TransparentSigningSet;
     use zcash_protocol::consensus;
 
     use super::{BuildResult, Builder, Error};
@@ -1123,8 +1327,8 @@ mod tests {
     use super::{Builder, Error};
     use crate::transaction::builder::BuildConfig;
 
-    use ::sapling::{Node, Rseed, zip32::ExtendedSpendingKey};
-    use ::transparent::{address::TransparentAddress, builder::TransparentSigningSet};
+    use sapling::{Node, Rseed, zip32::ExtendedSpendingKey};
+    use transparent::{address::TransparentAddress, builder::TransparentSigningSet};
     use zcash_protocol::{
         consensus::{NetworkUpgrade, Parameters, TEST_NETWORK},
         memo::MemoBytes,
@@ -1138,7 +1342,7 @@ mod tests {
     #[cfg(feature = "transparent-inputs")]
     use {
         crate::transaction::{OutPoint, TxOut, builder::DEFAULT_TX_EXPIRY_DELTA},
-        ::transparent::keys::{AccountPrivKey, IncomingViewingKey},
+        transparent::keys::{AccountPrivKey, IncomingViewingKey},
         zip32::AccountId,
     };
 
@@ -1148,7 +1352,7 @@ mod tests {
     #[cfg(feature = "transparent-inputs")]
     fn binding_sig_absent_if_no_shielded_spend_or_output() {
         use crate::transaction::builder::{self, TransparentBuilder};
-        use ::transparent::{builder::TransparentSigningSet, keys::NonHardenedChildIndex};
+        use transparent::{builder::TransparentSigningSet, keys::NonHardenedChildIndex};
         use zcash_protocol::consensus::NetworkUpgrade;
 
         let sapling_activation_height = TEST_NETWORK